js based quickfix
