+++ apps/autonolas-registry/common-util/Details/NFTImage/index.jsx
@@ -1,5 +1,5 @@
-import { useState, useEffect } from 'react';
 import PropTypes from 'prop-types';
+import { useEffect, useState } from 'react';
 
 import { useScreen } from '../../hooks/useScreen';
 import { NftImageContainer } from '../styles';
